+++ src/point.js
@@ -1,7 +1,7 @@
-import {createElement} from './utils.js';
-
-export class Point {
+import {Component} from './component.js';
+export class Point extends Component {
   constructor(data) {
+    super();
     this._type = data.type;
     this._icon = data.icon;
     this._city = data.city;
@@ -13,14 +13,12 @@ export class Point {
 
     this._element = null;
     this._onEdit = null;
+
+    this._onEditButtonClick = this._onEditButtonClick.bind(this);
   }
 
   _onEditButtonClick() {
     return typeof this._onEdit === `function` && this._onEdit();
-  }
-
-  get element() {
-    return this._element;
   }
 
   set onEdit(fn) {
@@ -43,30 +41,15 @@ export class Point {
     </article>`.trim();
   }
 
-  render() {
-    this._element = createElement(this.template);
-    this.bind();
-    return this._element;
-  }
-
   bind() {
     this._element
-        .addEventListener(`click`, this._onEditButtonClick.bind(this));
-  }
-
-  render() {
-    this._element = createElement(this.template);
-    this.bind();
-    return this._element;
+        .addEventListener(`click`, this._onEditButtonClick);
   }
 
   unbind() {
-    // Удаление обработчиков
+    this._element
+        .removeEvenetListener(`click`, this._onEditButtonClick);
   }
 
-  unrender() {
-    this.unbind();
-    this._element = null;
-  }
 }
